feat : Added create tag option in dropdown (#11296)
