+++ public/locale/en.json
@@ -828,6 +828,7 @@
   "create_schedule_template": "Create Schedule Template",
   "create_summaries": "Create Summaries",
   "create_tag": "Create Tag",
+  "create_tag_description": "Create a new tag for the questionnaire",
   "create_template": "Create Template",
   "create_user": "Create User",
   "create_user_and_add_to_org": "Create a new user and add them to the organization.",
@@ -1711,6 +1712,7 @@
   "my_profile": "My Profile",
   "my_schedules": "My Schedules",
   "name": "Name",
+  "name_and_slug_are_required": "Name and slug are required",
   "name_is_required": "Name is required",
   "name_of_hospital": "Name of Hospital",
   "name_of_shifting_approving_facility": "Name of shifting approving facility",
@@ -2630,8 +2632,10 @@
   "system": "System",
   "systolic": "Systolic",
   "tachycardia": "Tachycardia",
+  "tag_created_successfully": "Tag created successfully",
   "tag_name": "Tag Name",
   "tag_slug": "Tag Slug",
+  "tag_updated_successfully": "Tag updated successfully",
   "tags": "Tags",
   "taken": "Taken",
   "taper_titrate_dosage": "Taper & Titrate Dosage",
